requirements.txt
fastapi
uvicorn
streamlit
google-cloud-firestore
google-cloud-aiplatform
google-cloud-logging
google-api-python-client
google-auth-oauthlib
pydantic
google-cloud-aiplatform